Leo-Auto/requirements.txt
streamlit
pandas
google-generativeai
Pillow
fastapi
uvicorn